silly

+++ contextgen/detection/grid/join.py
@@ -2,7 +2,7 @@ from fire import Fire
 import os 
 from ..join_scores import interpolated_scores
 
-def grid_join(score_dir : str, rel_dir : str, out_dir : str, max_alpha : int):
+def grid_join(score_dir : str, rel_dir : str, out_dir : str, max_alpha : float):
     for alpha in range(0, max_alpha + 0.1, 0.1):
         alpha_name = str(alpha).replace('.', '_')
         out = os.path.join(out_dir, alpha_name)
